+++ nitrite-native-tests/src/test/java/org/dizitart/nitrite/test/repository/model/Title.java
@@ -1,0 +1,6 @@
+package org.dizitart.nitrite.test.repository.model;
+
+public enum Title {
+    DR,
+    PROF
+}