+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Issue Fixes
 
 - Fix for `Document.getFields()` not returning iterable fields
+- Fix for failing tests on systems with non-ENGLISH locale #994
 
 ## Release 4.3.0 - Jul 1, 2024